refactor(oauth): refactoring oauth
